Move related methods to Dao controller
