+++ packages/neuron-wallet/src/controllers/dao.ts
@@ -1,8 +1,9 @@
-import { Cell } from 'types/cell-types'
-import CellsService from '../services/cells'
-import { ServiceHasNoResponse } from 'exceptions'
+import { Cell, OutPoint, TransactionWithoutHash } from 'types/cell-types'
+import { ServiceHasNoResponse, IsRequired } from 'exceptions'
 import { ResponseCode } from 'utils/const'
+import WalletsService from 'services/wallets'
 import AddressesService from 'services/addresses'
+import CellsService from 'services/cells'
 import LockUtils from 'models/lock-utils'
 
 export default class DaoController {
@@ -22,6 +23,94 @@ export default class DaoController {
     return {
       status: ResponseCode.Success,
       result: cells,
+    }
+  }
+
+  public static async generateDepositTx(params: {
+    walletID: string,
+    capacity: string,
+    fee: string,
+    feeRate: string,
+  }): Promise<Controller.Response<TransactionWithoutHash>> {
+    if (!params) {
+      throw new IsRequired('Parameters')
+    }
+    const walletsService = WalletsService.getInstance()
+    const tx = await walletsService.generateDepositTx(
+      params.walletID,
+      params.capacity,
+      params.fee,
+      params.feeRate,
+    )
+    return {
+      status: ResponseCode.Success,
+      result: tx,
+    }
+  }
+
+  public static async generateDepositAllTx(params: {
+    walletID: string,
+    fee: string,
+    feeRate: string,
+  }): Promise<Controller.Response<TransactionWithoutHash>> {
+    if (!params) {
+      throw new IsRequired('Parameters')
+    }
+    const walletsService = WalletsService.getInstance()
+    const tx = await walletsService.generateDepositAllTx(
+      params.walletID,
+      params.fee,
+      params.feeRate,
+    )
+    return {
+      status: ResponseCode.Success,
+      result: tx,
+    }
+  }
+
+  public static async startWithdrawFromDao(params: {
+    walletID: string,
+    outPoint: OutPoint,
+    fee: string,
+    feeRate: string,
+  }): Promise<Controller.Response<TransactionWithoutHash>> {
+    if (!params) {
+      throw new IsRequired('Parameters')
+    }
+    const walletsService = WalletsService.getInstance()
+    const tx = await walletsService.startWithdrawFromDao(
+      params.walletID,
+      params.outPoint,
+      params.fee,
+      params.feeRate,
+    )
+    return {
+      status: ResponseCode.Success,
+      result: tx,
+    }
+  }
+
+  public static async withdrawFromDao(params: {
+    walletID: string,
+    depositOutPoint: OutPoint,
+    withdrawingOutPoint: OutPoint,
+    fee: string,
+    feeRate: string,
+  }): Promise<Controller.Response<TransactionWithoutHash>> {
+    if (!params) {
+      throw new IsRequired('Parameters')
+    }
+    const walletsService = WalletsService.getInstance()
+    const tx = await walletsService.withdrawFromDao(
+      params.walletID,
+      params.depositOutPoint,
+      params.withdrawingOutPoint,
+      params.fee,
+      params.feeRate,
+    )
+    return {
+      status: ResponseCode.Success,
+      result: tx,
     }
   }
 }
