+++ packages/neuron-wallet/src/controllers/wallets.ts
@@ -23,7 +23,7 @@ import {
 import i18n from 'utils/i18n'
 import AddressService from 'services/addresses'
 import WalletCreatedSubject from 'models/subjects/wallet-created-subject'
-import { TransactionWithoutHash, OutPoint } from 'types/cell-types'
+import { TransactionWithoutHash } from 'types/cell-types'
 import { MainnetAddressRequired, TestnetAddressRequired } from 'exceptions/address'
 
 export default class WalletsController {
@@ -392,94 +392,6 @@ export default class WalletsController {
     const tx = await walletsService.generateSendingAllTx(
       params.walletID,
       params.items,
-      params.fee,
-      params.feeRate,
-    )
-    return {
-      status: ResponseCode.Success,
-      result: tx,
-    }
-  }
-
-  public static async generateDepositTx(params: {
-    walletID: string,
-    capacity: string,
-    fee: string,
-    feeRate: string,
-  }): Promise<Controller.Response<TransactionWithoutHash>> {
-    if (!params) {
-      throw new IsRequired('Parameters')
-    }
-    const walletsService = WalletsService.getInstance()
-    const tx = await walletsService.generateDepositTx(
-      params.walletID,
-      params.capacity,
-      params.fee,
-      params.feeRate,
-    )
-    return {
-      status: ResponseCode.Success,
-      result: tx,
-    }
-  }
-
-  public static async startWithdrawFromDao(params: {
-    walletID: string,
-    outPoint: OutPoint,
-    fee: string,
-    feeRate: string,
-  }): Promise<Controller.Response<TransactionWithoutHash>> {
-    if (!params) {
-      throw new IsRequired('Parameters')
-    }
-    const walletsService = WalletsService.getInstance()
-    const tx = await walletsService.startWithdrawFromDao(
-      params.walletID,
-      params.outPoint,
-      params.fee,
-      params.feeRate,
-    )
-    return {
-      status: ResponseCode.Success,
-      result: tx,
-    }
-  }
-
-  public static async withdrawFromDao(params: {
-    walletID: string,
-    depositOutPoint: OutPoint,
-    withdrawingOutPoint: OutPoint,
-    fee: string,
-    feeRate: string,
-  }): Promise<Controller.Response<TransactionWithoutHash>> {
-    if (!params) {
-      throw new IsRequired('Parameters')
-    }
-    const walletsService = WalletsService.getInstance()
-    const tx = await walletsService.withdrawFromDao(
-      params.walletID,
-      params.depositOutPoint,
-      params.withdrawingOutPoint,
-      params.fee,
-      params.feeRate,
-    )
-    return {
-      status: ResponseCode.Success,
-      result: tx,
-    }
-  }
-
-  public static async generateDepositAllTx(params: {
-    walletID: string,
-    fee: string,
-    feeRate: string,
-  }): Promise<Controller.Response<TransactionWithoutHash>> {
-    if (!params) {
-      throw new IsRequired('Parameters')
-    }
-    const walletsService = WalletsService.getInstance()
-    const tx = await walletsService.generateDepositAllTx(
-      params.walletID,
       params.fee,
       params.feeRate,
     )
